feature: 작업자 배정 관련 필요기능 추가
- 교대유형 엔티티 레포 서비스 수정
- 작업지시 작업자배정 레포 서비스 수정 등

+++ src/main/java/com/megazone/ERPSystem_phase2_Backend/production/model/resource_data/dto/ShiftTypeDTO.java
@@ -13,5 +13,6 @@ public class ShiftTypeDTO {
     private Long id;
     private String name;       // 교대 근무 이름
     private String description; // 교대 근무 설명
-    private Long duration;     // 근무 시간 (시간 단위)
+    private Double duration;     // 근무 시간 (시간 단위)
+
 }
